requirements.txt
weaviate-client>=4.0.0
langchain
pydantic
